BaseApi: add helpers to handle server configurations
We do not want to use the m_serverConfigs and m_serverIndices maps
outside of the BaseApi class, so add helper methods to set server
configurations for operations and to set a base url for an operation
based on the selected server.

These methods are added as protected, because they're only supposed
to be used in the generated client API classes.

This also eliminates the last usage of m_networkFactory from the
api-body.mustache file.

This commit is a pre-requisite for moving the BaseApi member
variables into a private class.

Found in API review.

Task-number: QTBUG-142312
Pick-to: 6.11
Change-Id: Ia522f1c1c54cd88e03b346e0e0b2a42e5a84c4e9
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>