BaseApi: move member variables into a private class
This allows to further extend this class in future without breaking
binary compatibility.
Note that we do not derive the private class from QObjectPrivate,
because this code can be potentially compiled as a standalone project,
not as QtOpenApiCommon library, so we do not want it to depend on
Qt::CorePrivate library.

Found in API review.

Fixes: QTBUG-142312
Pick-to: 6.11
Change-Id: I6b36fde955c320f0a55dc4404180b6508e3852d4
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>