BaseApi: move member variables into a private class
This allows to further extend this class in future without breaking
binary compatibility.
Note that we do not derive the private class from QObjectPrivate,
because this code can be potentially compiled as a standalone project,
not as QtOpenApiCommon library, so we do not want it to depend on
Qt::CorePrivate library.

Found in API review.

Fixes: QTBUG-142312
Change-Id: I6b36fde955c320f0a55dc4404180b6508e3852d4
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>
(cherry picked from commit 2175b9ce9eef5931b7b135e267e58cc2616e127b)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>