Update the description of the useCmakeMacro option
The option is used to represent that the generator was called
from the qt_add_openapi_client macro.
That can enable all sorts of specific behaviors, not only the
generation of the CombinedModelsAndAPIs.cpp file.

Provide a more general description and advise against changing
this option manually.

This is a pre-requisite for the next commit that will add more
logic that depends on this option.

Task-number: QTBUG-142567
Change-Id: Idb17bdf426d77f8bc45d9918bf1bec035a4d5b98
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>
Reviewed-by: Rym Bouabid <rym.bouabid@qt.io>
(cherry picked from commit 385afd0d5870f4335eacffa5e7fb3d44ccbe7fb9)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>