Use QT_{BEGIN,END}_NAMESPACE when building Qt::OpenApiCommon
We need to use the proper Qt macros in order to make moc happy
when we're building the OpenApiCommon library as a part of Qt.
However, these macros should not be used when users builds all
the code on their own.

To solve the issue, we introduce a new "hidden" property into
the Java generator. This property is only set to true when the
qt_add_openapi_client() macro is used to build the common lib,
because this means that we're building the common lib as a part
of the Qt build.

Later this new property is used in the templates to generate the
code that is only available when building Qt.

This allows us to reintroduce the full namespace in the base
class specification in api-header.mustache, and thus get rid of
the qmltyperegistrar warning about a base type being used but not
found and other related qmllint warnings that were seen when
building the colorpalette example.

Fixes: QTBUG-142567
Pick-to: 6.11
Change-Id: I772d67edd90aed4683b1622582f39e3ab6494276
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>