Fix incorrect usage of OptionalParam in tests
The class actually belongs to QtOpenApiCommon namespace, not
QtOpenAPI.
It was working just by accident, most probably because we have
the "using namespace" declarations for both of the namespaces in
the beginning of each test.

Still, it was not correct, so fix it by simply not using any
namespace.

Change-Id: I13a01ea0154e2622e843a916d79e84dae05878c5
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>
(cherry picked from commit 0db4665db07d74695d92da1ccdbfb07dcd006fdb)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>